requirements.txt
openai
openai-agents
line-bot-sdk
yfinance
beautifulsoup4
lxml
cloudscraper
stockstats
fastapi
uvicorn[standard]
python-multipart
python-dotenv